cleared backend issue

+++ controllers/todoController.js
@@ -1,27 +1,34 @@
 import Todo from '../models/Todo.js';
 
 export const getTodos = async (req, res) => {
-  const todos = await Todo.find({ user: req.user.id });
+  const todos = await Todo.find({ user: req.userId });
   res.json(todos);
 };
 
 export const createTodo = async (req, res) => {
-  const { text } = req.body;
-  const newTodo = await Todo.create({ text, user: req.user.id });
-  res.status(201).json(newTodo);
+  const todo = await Todo.create({ ...req.body, user: req.userId });
+  res.status(201).json(todo);
 };
 
 export const updateTodo = async (req, res) => {
-  const updated = await Todo.findByIdAndUpdate(req.params.id, req.body, { new: true });
+  const updated = await Todo.findByIdAndUpdate(
+    { _id: req.params.id, user: req.userId},
+    req.body,
+    { new: true }
+  );
   res.json(updated);
 };
 
 export const deleteTodo = async (req, res) => {
-  await Todo.findByIdAndDelete(req.params.id);
+  await Todo.findByIdAndDelete({ _id: req.params.id, user: req.userId });
   res.json({ message: 'Deleted' });
 };
 
-export const markTodoAsRead = async (req, res) => {
-  const updated = await Todo.findByIdAndUpdate(req.params.id, { read: true }, { new: true });
+export const markAsRead = async (req, res) => {
+  const updated = await Todo.findOneAndUpdate(
+    { _id: req.params.id, user: req.userId },
+    { read: true },
+    { new: true }
+  );
   res.json(updated);
 };
